fix: onboarding wizard improvements
